Fixed compile errors from long type clash

+++ src/test/java/io/craigmiller160/jaxrs/oauth2/domain/repository/JdbcAppRefreshTokenRepositoryTest.kt
@@ -71,7 +71,7 @@ class JdbcAppRefreshTokenRepositoryTest {
             } }
         }
 
-        repo.deleteById(id)
+        repo.deleteById(id as java.lang.Long)
 
         DriverManager.getConnection(getJdbcUrl()).use { conn ->
             conn.createStatement().use { stmt ->
